requirements.txt
telebot==0.0.4
pyTelegramBotAPI==4.2.2
requests
python-telegram-bot==13.8.1